Clean up config related codes to maintain consistency
- Extract rpc.Config and backend.Config into separate files
- Move Config Validation to yorkie.New
- Reorder the functions of yorkie.Config
