+++ yorkie/backend/backend.go
@@ -17,7 +17,6 @@
 package backend
 
 import (
-	"fmt"
 	"os"
 	gosync "sync"
 	"time"
@@ -26,7 +25,6 @@ import (
 
 	"github.com/yorkie-team/yorkie/internal/log"
 	"github.com/yorkie-team/yorkie/pkg/cache"
-	"github.com/yorkie-team/yorkie/pkg/types"
 	"github.com/yorkie-team/yorkie/yorkie/backend/db"
 	"github.com/yorkie-team/yorkie/yorkie/backend/db/mongo"
 	"github.com/yorkie-team/yorkie/yorkie/backend/sync"
@@ -36,66 +34,6 @@ import (
 )
 
 const authWebhookCacheSize = 5000
-
-// Config is the configuration for creating a Backend instance.
-type Config struct {
-	// SnapshotThreshold is the threshold that determines if changes should be
-	// sent with snapshot when the number of changes is greater than this value.
-	SnapshotThreshold uint64 `json:"SnapshotThreshold"`
-
-	// SnapshotInterval is the interval of changes to create a snapshot.
-	SnapshotInterval uint64 `json:"SnapshotInterval"`
-
-	// AuthorizationWebhookURL is the url of the authorization webhook.
-	AuthorizationWebhookURL string `json:"AuthorizationWebhookURL"`
-
-	// AuthorizationWebhookMethods is the methods that run the authorization webhook.
-	AuthorizationWebhookMethods []string `json:"AuthorizationWebhookMethods"`
-
-	// AuthorizationWebhookMaxRetries is the max count
-	// that retries the authorization webhook.
-	AuthorizationWebhookMaxRetries uint64 `json:"AuthorizationWebhookMaxRetries"`
-
-	// AuthorizationWebhookMaxWaitIntervalMillis is the max interval
-	// that waits before retrying the authorization webhook.
-	AuthorizationWebhookMaxWaitIntervalMillis uint64 `json:"AuthorizationWebhookMaxWaitIntervalMillis"`
-
-	// AuthorizationWebhookCacheAuthorizedTTLSec is the TTL value to set when caching the authorized result.
-	AuthorizationWebhookCacheAuthorizedTTLSec uint64 `json:"AuthorizationWebhookCacheAuthorizedTTLSec"`
-
-	// AuthorizationWebhookCacheAuthorizedTTLSec is the TTL value to set when caching the unauthorized result.
-	AuthorizationWebhookCacheUnauthorizedTTLSec uint64 `json:"AuthorizationWebhookCacheUnauthorizedTTLSec"`
-}
-
-// RequireAuth returns whether the given method require authorization.
-func (c *Config) RequireAuth(method types.Method) bool {
-	if len(c.AuthorizationWebhookURL) == 0 {
-		return false
-	}
-
-	if len(c.AuthorizationWebhookMethods) == 0 {
-		return true
-	}
-
-	for _, m := range c.AuthorizationWebhookMethods {
-		if types.Method(m) == method {
-			return true
-		}
-	}
-
-	return false
-}
-
-// Validate validates this config.
-func (c *Config) Validate() error {
-	for _, method := range c.AuthorizationWebhookMethods {
-		if !types.IsAuthMethod(method) {
-			return fmt.Errorf("not supported method for authorization webhook: %s", method)
-		}
-	}
-
-	return nil
-}
 
 // Backend manages Yorkie's backend such as Database and Coordinator. And it
 // has the server status such as the information of this Agent.
