+++ yorkie/rpc/server.go
@@ -18,10 +18,8 @@ package rpc
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"net"
-	"os"
 
 	grpcmiddleware "github.com/grpc-ecosystem/go-grpc-middleware"
 	grpcprometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
@@ -35,22 +33,6 @@ import (
 	"github.com/yorkie-team/yorkie/yorkie/backend"
 	"github.com/yorkie-team/yorkie/yorkie/rpc/interceptors"
 )
-
-var (
-	// ErrInvalidRPCPort occurs when the port in the config is invalid.
-	ErrInvalidRPCPort = errors.New("invalid port number for RPC server")
-	// ErrInvalidCertFile occurs when the certificate file is invalid.
-	ErrInvalidCertFile = errors.New("invalid cert file for RPC server")
-	// ErrInvalidKeyFile occurs when the key file is invalid.
-	ErrInvalidKeyFile = errors.New("invalid key file for RPC server")
-)
-
-// Config is the configuration for creating a Server instance.
-type Config struct {
-	Port     int
-	CertFile string
-	KeyFile  string
-}
 
 // Server is a normal server that processes the logic requested by the client.
 type Server struct {
@@ -133,28 +115,6 @@ func (s *Server) listenAndServeGRPC() error {
 			}
 		}
 	}()
-
-	return nil
-}
-
-// Validate validates the port number and the files for certification.
-func (c *Config) Validate() error {
-	if c.Port < 1 || 65535 < c.Port {
-		return fmt.Errorf("must be between 1 and 65535, given %d: %w", c.Port, ErrInvalidRPCPort)
-	}
-
-	// when specific cert or key file are configured
-	if c.CertFile != "" {
-		if _, err := os.Stat(c.CertFile); err != nil {
-			return fmt.Errorf("%s: %w", c.CertFile, ErrInvalidCertFile)
-		}
-	}
-
-	if c.KeyFile != "" {
-		if _, err := os.Stat(c.KeyFile); err != nil {
-			return fmt.Errorf("%s: %w", c.KeyFile, ErrInvalidKeyFile)
-		}
-	}
 
 	return nil
 }
